Add files via upload
Changes for revision M

BIG things:
1. Use EEPROM to store centering and scaling information;
Calcualte and store the dcXoffset and dcYoffset values and the four extent 
values for the SC scale command. This results in the printed plot in 1:1 scale
(10cmx10cm). With pin 7 low use the XYSetup function of the scope. The values 
will be stored in EEPROM and recalled for all future plots.
(Make pin 7 high when finished with XYSetup!)

Little things:
1. Put some functionaltiy in their owm methods
  a. Go_NoGo() for red/gree LED control
  b. waitForNominalVoltage() for voltage monitoring
  c. inline IdleLED_(x)
2. Use signed integers since dcOffsets can be negative